Just a temporary push, I will be disabling the radom spell gain from this POI before we  release, it is literally just here for now.
The goal is to give the player the dragon training interaction when they first arrive in a valyria province.

When I write them, the blackstone interactions will also be acquired like this, by a random chance when you first enter Valyria provinces. Will add a triggered event, but for now it is just to give you the dragon training action.